test: split test utils into dedicated files
