+++ internal/testutils/testserver.go
@@ -10,7 +10,6 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"regexp"
 	"strings"
 	"testing"
 	"time"
@@ -416,24 +415,4 @@ func (ts *testServer) MustRegister(t *testing.T, queries *mocks.MockDBQueries, u
 
 	postResult := ts.PostForm(t, "/user/register", form)
 	assert.Equal(t, http.StatusSeeOther, postResult.Status)
-}
-
-var rxCSRF = regexp.MustCompile(`<input type="hidden" name="csrf_token" value="(.+?)">`)
-
-func ExtractCSRFToken(t *testing.T, body string) string {
-	matches := rxCSRF.FindStringSubmatch(body)
-	if len(matches) < 2 {
-		t.Fatalf("no csrf token found in body")
-	}
-
-	return matches[1]
-}
-
-func MustHashPassword(t *testing.T, hasher crypto.PasswordHasher, password string) string {
-	t.Helper()
-	hash, err := hasher.GenerateFromPassword(password, crypto.DefaultPasswordHashParams)
-	if err != nil {
-		t.Fatalf("error generating hash for password %q: %v", password, err)
-	}
-	return hash
 }
